+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - View for viewing event group details
 - Breadcrumbs to event and event group views
+- Event group creation view
 
 ### Changed